[Cache] Put the internal ClusterQueue on a separate file
