+++ pkg/cache/cache_test.go
@@ -32,7 +32,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	kueue "sigs.k8s.io/kueue/apis/kueue/v1beta1"
-	"sigs.k8s.io/kueue/pkg/metrics"
 	"sigs.k8s.io/kueue/pkg/util/pointer"
 	utiltesting "sigs.k8s.io/kueue/pkg/util/testing"
 	"sigs.k8s.io/kueue/pkg/workload"
@@ -2108,65 +2107,6 @@ func TestClusterQueuesUsingFlavor(t *testing.T) {
 				return a < b
 			})); len(diff) != 0 {
 				t.Errorf("Unexpected flavor is in use by clusterQueues (-want,+got):\n%s", diff)
-			}
-		})
-	}
-}
-
-func TestClusterQueueUpdateWithFlavors(t *testing.T) {
-	rf := utiltesting.MakeResourceFlavor("x86").Obj()
-	cq := utiltesting.MakeClusterQueue("cq").
-		ResourceGroup(*utiltesting.MakeFlavorQuotas("x86").Resource("cpu", "5").Obj()).
-		Obj()
-
-	testcases := []struct {
-		name       string
-		curStatus  metrics.ClusterQueueStatus
-		flavors    map[kueue.ResourceFlavorReference]*kueue.ResourceFlavor
-		wantStatus metrics.ClusterQueueStatus
-	}{
-		{
-			name:      "Pending clusterQueue updated existent flavors",
-			curStatus: pending,
-			flavors: map[kueue.ResourceFlavorReference]*kueue.ResourceFlavor{
-				kueue.ResourceFlavorReference(rf.Name): rf,
-			},
-			wantStatus: active,
-		},
-		{
-			name:       "Active clusterQueue updated with not found flavors",
-			curStatus:  active,
-			flavors:    map[kueue.ResourceFlavorReference]*kueue.ResourceFlavor{},
-			wantStatus: pending,
-		},
-		{
-			name:      "Terminating clusterQueue updated with existent flavors",
-			curStatus: terminating,
-			flavors: map[kueue.ResourceFlavorReference]*kueue.ResourceFlavor{
-				kueue.ResourceFlavorReference(rf.Name): rf,
-			},
-			wantStatus: terminating,
-		},
-		{
-			name:       "Terminating clusterQueue updated with not found flavors",
-			curStatus:  terminating,
-			wantStatus: terminating,
-		},
-	}
-
-	for _, tc := range testcases {
-		t.Run(tc.name, func(t *testing.T) {
-			cache := New(utiltesting.NewFakeClient())
-			cq, err := cache.newClusterQueue(cq)
-			if err != nil {
-				t.Fatalf("failed to new clusterQueue %v", err)
-			}
-
-			cq.Status = tc.curStatus
-			cq.UpdateWithFlavors(tc.flavors)
-
-			if cq.Status != tc.wantStatus {
-				t.Fatalf("got different status, want: %v, got: %v", tc.wantStatus, cq.Status)
 			}
 		})
 	}
