+++ pkg/cache/clusterqueue_test.go
@@ -1,0 +1,68 @@
+package cache
+
+import (
+	"testing"
+
+	kueue "sigs.k8s.io/kueue/apis/kueue/v1beta1"
+	"sigs.k8s.io/kueue/pkg/metrics"
+	utiltesting "sigs.k8s.io/kueue/pkg/util/testing"
+)
+
+func TestClusterQueueUpdateWithFlavors(t *testing.T) {
+	rf := utiltesting.MakeResourceFlavor("x86").Obj()
+	cq := utiltesting.MakeClusterQueue("cq").
+		ResourceGroup(*utiltesting.MakeFlavorQuotas("x86").Resource("cpu", "5").Obj()).
+		Obj()
+
+	testcases := []struct {
+		name       string
+		curStatus  metrics.ClusterQueueStatus
+		flavors    map[kueue.ResourceFlavorReference]*kueue.ResourceFlavor
+		wantStatus metrics.ClusterQueueStatus
+	}{
+		{
+			name:      "Pending clusterQueue updated existent flavors",
+			curStatus: pending,
+			flavors: map[kueue.ResourceFlavorReference]*kueue.ResourceFlavor{
+				kueue.ResourceFlavorReference(rf.Name): rf,
+			},
+			wantStatus: active,
+		},
+		{
+			name:       "Active clusterQueue updated with not found flavors",
+			curStatus:  active,
+			flavors:    map[kueue.ResourceFlavorReference]*kueue.ResourceFlavor{},
+			wantStatus: pending,
+		},
+		{
+			name:      "Terminating clusterQueue updated with existent flavors",
+			curStatus: terminating,
+			flavors: map[kueue.ResourceFlavorReference]*kueue.ResourceFlavor{
+				kueue.ResourceFlavorReference(rf.Name): rf,
+			},
+			wantStatus: terminating,
+		},
+		{
+			name:       "Terminating clusterQueue updated with not found flavors",
+			curStatus:  terminating,
+			wantStatus: terminating,
+		},
+	}
+
+	for _, tc := range testcases {
+		t.Run(tc.name, func(t *testing.T) {
+			cache := New(utiltesting.NewFakeClient())
+			cq, err := cache.newClusterQueue(cq)
+			if err != nil {
+				t.Fatalf("failed to new clusterQueue %v", err)
+			}
+
+			cq.Status = tc.curStatus
+			cq.UpdateWithFlavors(tc.flavors)
+
+			if cq.Status != tc.wantStatus {
+				t.Fatalf("got different status, want: %v, got: %v", tc.wantStatus, cq.Status)
+			}
+		})
+	}
+}
